Restructing readme, combining elements into one descriptive doc.
