+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"fmt"
 	"time"
 	"workpath/delegator"
 	"workpath/worker"
@@ -11,7 +12,9 @@ func main() {
 	w := worker.NewJob(HelloWorld, nil)
 	for {
 		w.Run(dchan)
+		fmt.Println("Job sent to worker")
 		<-dchan
+		fmt.Println("Job completed")
 		time.Sleep(3 * time.Second)
 	}
 }
