+++ workpath/worker/job.go
@@ -1,56 +1,113 @@
 package worker
 
-import "sync"
+// Worker is a package designed to execute a task in a goroutine,
+// with a variable map and task specific metrics. It is
+// designed to be used with the delegator package. The task
+// is defined as an enum, which allows for easy addition of new tasks.
+import (
+	"fmt"
+	"sync"
+	"workpath/delegator"
 
-type work struct {
-	task    task   // The task to be performed.
-	done    bool   // Indicates if the task has been run.
-	payload []byte // The returned value of the task.
-	e       error  // the error returned by the task.
-	vars    VMap   // Variables used by the task.
-}
+	"github.com/google/uuid"
 
-type (
-	// The task Signature expected by the worker.
-	TaskSignature func(VMap) (error, []byte)
+	"log"
 )
 
-// Set work payload.
-func (w *work) SetPayload(p []byte) {
-	w.payload = p
-}
+// Define enums used to indicate state of the job.
+type state int64
 
-// Get work payload.
-func (w *work) GetPayload() []byte {
-	return w.payload
-}
+const (
+	running state = iota
+	completed
+	failed
+)
 
-// Return Error value of the job.
-func (w *work) Error() error {
-	return w.e
-}
-
-func (w *work) IsDone() bool {
-	return w.done
-}
-
-// Returns the "variable map" used by the job.
-// A pointer is returned, modify the map via
-// the SetKeys, Get, and Delete methods.
-func (w *work) GetVmap() *VMap {
-	return &w.vars
-}
-
-func (w *work) execute(wg *sync.WaitGroup) {
-	defer func() {
-		w.done = true
-		wg.Done()
-	}()
-	err, payload := w.task.Func()(w.vars)
-	w.SetPayload(payload)
-	if err != nil {
-		w.e = err
-		return
+func (a state) String() string {
+	switch a {
+	case running:
+		return "still running"
+	case completed:
+		return "completed successfully"
+	case failed:
+		return "failed"
 	}
-	w.e = nil
+	return ""
+}
+
+type job struct {
+	id   uint32 // Unique id of the job
+	work work   // The work to be done.
+}
+
+// Worker interface represents a job that can be executed.
+type Worker interface {
+	GetState() state
+	ID() uint32
+	Run(done chan delegator.Directive)
+	logError()
+}
+
+// return the state of the job.
+func (j *job) GetState() state {
+	w := &j.work
+	if w.IsDone() {
+		if w.Error() != nil {
+			return failed
+		}
+		return running
+	}
+	return completed
+}
+
+// Returns the unique id of the job.
+func (j *job) ID() uint32 {
+	return j.id
+}
+
+// Runs the job in a goroutine.
+func (j *job) Run(done chan delegator.Directive) {
+	w := &j.work
+	var wg sync.WaitGroup
+	defer func() {
+		j.Announce()
+		fmt.Println("waiting to send completion message to delegator")
+		done <- delegator.NewDoneDirective(j.id)
+		fmt.Println("passed to delegator")
+	}()
+	wg.Add(1)
+	func() { go w.execute(); wg.Done() }()
+	wg.Wait()
+}
+
+// handle any errors returned by the task.
+// If the error is not nil, it is logged.
+func (j *job) logError() {
+	if j.work.Error() != nil {
+		log.Println(j.work.Error())
+	}
+}
+
+// Creates a new, nil variable map.
+func NewVmap() *VMap {
+	return &VMap{}
+}
+
+// constructs a new job.
+func NewJob(t task, v map[string]any) *job {
+	w := work{
+		task: t,
+		vars: v,
+		done: false,
+	}
+	return &job{
+		id:   uuid.New().ID(),
+		work: w,
+	}
+}
+
+func (j *job) Announce() {
+	// Log the completion of the job.
+	log.Printf("Job %d %v", j.id, j.GetState().String())
+	j.logError()
 }
